+++ HISTORY.rst
@@ -97,6 +97,7 @@ History
 - Added default and configurable list of allowed languages and countries
 - Added `VC-1` as new `video_codec` possible value
 - Enhanced dash-separated `release_group` detection.
+- Changed `size` output to return `guessit.Quantity` object.
 
 2.1.4 (2017-06-01)
 ------------------